+++ front/src/form/bsdasri/Transporter.tsx
@@ -179,7 +179,7 @@ function BaseTransporter({ status, displayTakeoverFields = false, stepName }) {
         <div className="form__row">
           <label>
             Immatriculations
-            <Tooltip msg="Saisissez les numéros un par un. Appuyez sur la touche <Entrée> pour valider chacun" />
+            <Tooltip msg="Saisissez les numéros un par un. Appuyez sur la touche <Entrée> ou <Tab> pour valider chacun" />
             <TagsInput
               name="transporter.transport.plates"
               disabled={disabled}
@@ -187,6 +187,7 @@ function BaseTransporter({ status, displayTakeoverFields = false, stepName }) {
           </label>
         </div>
       )}
+
       <div
         className={classNames("form__row", {
           "field-emphasis": transportEmphasis,
@@ -258,6 +259,13 @@ function BaseTransporter({ status, displayTakeoverFields = false, stepName }) {
           l'emport du déchet
         </p>
       )}
+      <div className="form__row">
+        <label>
+          Numéros de containers
+          <Tooltip msg="Saisissez les numéros un par un. Appuyez sur la touche <Entrée> ou <Tab> pour valider chacun" />
+          <TagsInput name="identification.numbers" disabled={disabled} />
+        </label>
+      </div>
     </>
   );
 }